+++ nengo/conftest.py
@@ -1,9 +1,10 @@
+from fnmatch import fnmatch
 import hashlib
 import inspect
 import importlib
 import os
 import re
-from fnmatch import fnmatch
+import shlex
 import warnings
 
 import matplotlib
@@ -386,29 +387,34 @@ def pytest_report_collectionfinish(config, startdir, items):
     return deselect_reasons
 
 
-def pytest_runtest_setup(item):  # noqa: C901
+def pytest_runtest_setup(item):
     rc.reload_rc([])
     rc.set('decoder_cache', 'enabled', 'False')
     rc.set('exceptions', 'simplified', 'False')
 
-    if not hasattr(item, 'obj'):
-        return  # Occurs for doctests, possibly other weird tests
+    item_name = get_item_name(item)
 
-    test_uses_sim = 'Simulator' in item.fixturenames
-    test_uses_refsim = 'RefSimulator' in item.fixturenames
-    tests_frontend = not (test_uses_sim or test_uses_refsim)
+    # join all the lines and then split (preserving quoted strings)
+    unsupported = shlex.split(
+        " ".join(item.config.getini("nengo_test_unsupported")))
+    # group pairs (representing testname + reason)
+    unsupported = [
+        unsupported[i:i + 2] for i in range(0, len(unsupported), 2)]
 
-    if not tests_frontend:
-        item_name = get_item_name(item)
+    for test, reason in unsupported:
+        # wrap square brackets to interpret them literally
+        # (see https://docs.python.org/3/library/fnmatch.html)
+        test = "".join("[%s]" % c if c in ('[', ']') else c for c in test)
 
-        for test, reason in TestConfig.Simulator.unsupported:
-            # We add a '*' before test to eliminate the surprise of needing
-            # a '*' before the name of a test function.
-            if fnmatch(item_name, '*' + test):
-                if TestConfig.run_unsupported:
-                    item.add_marker(pytest.mark.xfail)
-                else:
-                    pytest.skip(reason)
+        # We add a '*' before test to eliminate the surprise of needing
+        # a '*' before the name of a test function.
+        test = "*" + test
+
+        if fnmatch(item_name, test):
+            if TestConfig.run_unsupported:
+                item.add_marker(pytest.mark.xfail(reason=reason))
+            else:
+                pytest.skip(reason)
 
 
 def pytest_terminal_summary(terminalreporter):